added new module for renderComments
